Fix circular dependency in tRPC setup
- Extract validateCliToken to standalone auth utility (src/server/auth/cli-token.ts)
- Break circular import between create-router.ts and cli-auth-router.ts
- Consolidate tRPC client to single module in src/utils/trpc.ts
- Fix createTRPCReact API usage (remove invalid {} parameter)
- Update all tRPC imports to use consolidated client

This resolves the "ReferenceError: Cannot access 'm' before initialization" build error.
