+++ packages/web/src/utils/trpc.ts
@@ -2,7 +2,7 @@ import { createTRPCReact } from '@trpc/react-query';
 import { inferRouterInputs, inferRouterOutputs } from '@trpc/server';
 import type { AppRouter } from '../server/routers/_app';
 
-export const trpc = createTRPCReact<AppRouter>({});
+export const trpc = createTRPCReact<AppRouter>();
 
 /**
  * Inference helpers for input types
